requirements.txt
Flask==2.3.2
python-telegram-bot==20.2
requests==2.28.1
gunicorn==20.1.0
